Define joingroup logics
